tests/test_config.py
"""
Test configuration system for V2
"""

import os
import tempfile
from pathlib import Path

import pytest
import yaml

# Placeholder for future config implementation
# from obsidian_processor.config import Config, ConfigError


class TestConfig:
    """Test configuration loading and validation"""

    def test_config_placeholder(self):
        """Placeholder test - will be updated when config.py is implemented"""
        # This is a placeholder test that should pass
        assert True

    def test_load_yaml_config(self, config_file: Path):
        """Test loading YAML configuration"""
        with open(config_file) as f:
            config = yaml.safe_load(f)

        assert config["vault_path"] == "/tmp/test_vault_fixture"
        assert "exclude_patterns" in config
        assert "processing" in config
        assert "processors" in config

    def test_config_validation(self, sample_config: dict):
        """Test configuration validation"""
        # Test required fields
        assert "vault_path" in sample_config
        assert "exclude_patterns" in sample_config
        assert "processing" in sample_config
        assert "processors" in sample_config

        # Test processing config
        processing = sample_config["processing"]
        assert processing["concurrency_limit"] > 0
        assert processing["retry_attempts"] >= 0
        assert processing["timeout"] > 0

    def test_environment_variable_interpolation(self, temp_dir: Path):
        """Test environment variable interpolation in config"""
        config_content = {
            "vault_path": "${TEST_VAULT_PATH}",
            "processors": {"test": {"type": "script", "config": {"api_key": "${TEST_API_KEY}"}}},
        }

        config_path = temp_dir / "test_config.yaml"
        with open(config_path, "w") as f:
            yaml.dump(config_content, f)

        # Set environment variables
        os.environ["TEST_VAULT_PATH"] = "/test/vault"
        os.environ["TEST_API_KEY"] = "test_key"

        try:
            # Test that config contains environment variables
            with open(config_path) as f:
                config = yaml.safe_load(f)

            assert config["vault_path"] == "${TEST_VAULT_PATH}"
            assert config["processors"]["test"]["config"]["api_key"] == "${TEST_API_KEY}"

        finally:
            # Clean up environment variables
            os.environ.pop("TEST_VAULT_PATH", None)
            os.environ.pop("TEST_API_KEY", None)

    def test_invalid_config_format(self, temp_dir: Path):
        """Test handling of invalid config format"""
        config_path = temp_dir / "invalid_config.yaml"
        config_path.write_text("invalid: yaml: content: [")

        with pytest.raises(yaml.YAMLError):
            with open(config_path) as f:
                yaml.safe_load(f)

    def test_missing_required_fields(self, temp_dir: Path):
        """Test validation of missing required fields"""
        incomplete_config = {"processing": {"concurrency_limit": 5}}

        config_path = temp_dir / "incomplete_config.yaml"
        with open(config_path, "w") as f:
            yaml.dump(incomplete_config, f)

        with open(config_path) as f:
            config = yaml.safe_load(f)

        # These fields should be missing
        assert "vault_path" not in config
        assert "exclude_patterns" not in config
        assert "processors" not in config

    def test_default_values(self, sample_config: dict):
        """Test default configuration values"""
        processing = sample_config["processing"]

        # Test reasonable defaults
        assert processing["concurrency_limit"] == 2
        assert processing["retry_attempts"] == 3
        assert processing["retry_delay"] == 1.0
        assert processing["timeout"] == 60

    def test_exclude_patterns_validation(self, sample_config: dict):
        """Test exclude patterns validation"""
        patterns = sample_config["exclude_patterns"]

        # Test that patterns are strings
        assert all(isinstance(p, str) for p in patterns)

        # Test expected patterns
        assert "templates/**" in patterns
        assert "archive/**" in patterns
        assert "**/*.template.md" in patterns
        assert "**/.*" in patterns

    def test_processor_config_validation(self, sample_config: dict):
        """Test processor configuration validation"""
        processors = sample_config["processors"]

        assert "test_processor" in processors

        processor = processors["test_processor"]
        assert "type" in processor
        assert "config" in processor
        assert processor["type"] == "script"
