requirements.txt
pyyaml>=6.0.0
aiofiles>=23.0.0
click>=8.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-anyio>=0.4.0
